web/requirements.txt
flask
requests-oauthlib
psycopg2-binary
pyyaml